+++ src/server.cpp
@@ -836,6 +836,10 @@ int Loop::client_request(ISlice name, Connect *client) {
         }
         if(server->log & 32) worker->start_time = get_time();
     } else {
+        if(client->header_option == "if present") {
+            client->send.status("204 No Content")->done(1);
+            return 0;
+        }
         bool inserted = false;
         auto *clients = &ql->queue[_nloop].clients;